coding 2016/03/21

+++ src/main/java/cn/edu/nuc/acmicpc/service/ProblemService.java
@@ -28,7 +28,7 @@ public interface ProblemService {
      * @param problemId
      * @return
      */
-    public ProblemDto getProblemDtoByProblemId(Integer problemId);
+    public ProblemDto getProblemDtoByProblemId(Long problemId);
 
     /**
      * Counts the number of problems fit in condition
